Refactored MaintenanceActions to use a seperate folder.
