+++ arangod/Cluster/MaintenanceActions/Action.cpp
@@ -25,18 +25,18 @@
 #include "Action.h"
 
 #include "Basics/Exceptions.h"
-#include "Cluster/CreateCollection.h"
-#include "Cluster/CreateDatabase.h"
-#include "Cluster/DropCollection.h"
-#include "Cluster/DropDatabase.h"
-#include "Cluster/DropIndex.h"
-#include "Cluster/EnsureIndex.h"
+#include "Cluster/MaintenanceActions/CreateCollection.h"
+#include "Cluster/MaintenanceActions/CreateDatabase.h"
+#include "Cluster/MaintenanceActions/DropCollection.h"
+#include "Cluster/MaintenanceActions/DropDatabase.h"
+#include "Cluster/MaintenanceActions/DropIndex.h"
+#include "Cluster/MaintenanceActions/EnsureIndex.h"
 #include "Cluster/MaintenanceStrings.h"
-#include "Cluster/ResignShardLeadership.h"
-#include "Cluster/SynchronizeShard.h"
-#include "Cluster/TakeoverShardLeadership.h"
-#include "Cluster/UpdateCollection.h"
-#include "Cluster/UpdateReplicatedLogAction.h"
+#include "Cluster/MaintenanceActions/ResignShardLeadership.h"
+#include "Cluster/MaintenanceActions/SynchronizeShard.h"
+#include "Cluster/MaintenanceActions/TakeoverShardLeadership.h"
+#include "Cluster/MaintenanceActions/UpdateCollection.h"
+#include "Cluster/MaintenanceActions/UpdateReplicatedLogAction.h"
 
 using namespace arangodb;
 using namespace arangodb::maintenance;
