+++ arangod/Cluster/MaintenanceActions/DropCollection.cpp
