+++ arangod/Cluster/MaintenanceActions/DropCollection.h
