+++ arangod/Cluster/MaintenanceActions/DropDatabase.cpp
