+++ arangod/Cluster/MaintenanceActions/DropIndex.cpp
