+++ arangod/Cluster/MaintenanceActions/ResignShardLeadership.cpp
