+++ arangod/Cluster/MaintenanceActions/ResignShardLeadership.h
