+++ arangod/Cluster/MaintenanceFeature.cpp
@@ -42,12 +42,12 @@
 #include "Basics/TimeString.h"
 #include "Basics/WriteLocker.h"
 #include "Basics/system-functions.h"
-#include "Cluster/Action.h"
-#include "Cluster/ActionDescription.h"
+#include "Cluster/MaintenanceActions/Action.h"
+#include "Cluster/MaintenanceActions/ActionDescription.h"
 #include "Cluster/AgencyCache.h"
 #include "Cluster/ClusterFeature.h"
 #include "Cluster/ClusterInfo.h"
-#include "Cluster/CreateDatabase.h"
+#include "Cluster/MaintenanceActions/CreateDatabase.h"
 #include "Cluster/MaintenanceWorker.h"
 #include "Cluster/ServerState.h"
 #include "Logger/LogMacros.h"
