+++ tests/Maintenance/MaintenanceTest.cpp
@@ -35,7 +35,7 @@
 #include "Basics/StaticStrings.h"
 #include "Cluster/Maintenance.h"
 #include "Cluster/MaintenanceFeature.h"
-#include "Cluster/ResignShardLeadership.h"
+#include "Cluster/MaintenanceActions/ResignShardLeadership.h"
 #include "Metrics/MetricsFeature.h"
 #include "Mocks/Servers.h"
 #include "Mocks/StorageEngineMock.h"
