+++ tests/Mocks/Servers.cpp
@@ -40,13 +40,13 @@
 #include "Basics/StringUtils.h"
 #include "Basics/TimeString.h"
 #include "Basics/files.h"
-#include "Cluster/ActionDescription.h"
+#include "Cluster/MaintenanceActions/ActionDescription.h"
 #include "Cluster/AgencyCache.h"
 #include "Cluster/ClusterFeature.h"
 #include "Cluster/ClusterInfo.h"
-#include "Cluster/CreateCollection.h"
-#include "Cluster/CreateDatabase.h"
-#include "Cluster/DropDatabase.h"
+#include "Cluster/MaintenanceActions/CreateCollection.h"
+#include "Cluster/MaintenanceActions/CreateDatabase.h"
+#include "Cluster/MaintenanceActions/DropDatabase.h"
 #include "Cluster/Maintenance.h"
 #include "ClusterEngine/ClusterEngine.h"
 #include "FeaturePhases/AqlFeaturePhase.h"
